Use policy registry
Use policy registry for HTTP requests rather than create per-request.
